requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
arrow==1.1.1; python_version >= '3.6'
asgiref==3.4.1; python_version >= '3.6'
astroid==2.2.5
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
awsebcli==3.20.1
backports.zoneinfo==0.2.1; python_version < '3.9'
bcrypt==3.2.0; python_version >= '3.6'
blessed==1.18.1; python_version >= '2.7'
botocore==1.21.33; python_version >= '3.6'
brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.2; python_version ~= '3.5'
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4; python_version >= '3'
click==8.0.1; python_version >= '3.6'
colorama==0.4.3; sys_platform != 'fake'
comtypes==1.1.10
cryptography==3.4.8; python_version >= '3.6'
cssselect2==0.4.1; python_version >= '3.6'
csvimporter==0.1.3.2
dateparser==1.0.0; python_version >= '3.5'
dj-database-url==0.5.0
django-compression-middleware==0.4.1
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-robots-txt==0.4
django-storages-azure==1.6.8
django==3.2.7
djangorestframework==3.12.4
docker-compose==1.25.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
docker[ssh]==4.4.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
google-api-core==1.31.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth==1.35.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
googleapis-common-protos==1.53.0; python_version >= '3.6'
gunicorn==20.1.0
h11==0.12.0; python_version >= '3.6'
idna==3.2; python_version >= '3'
isort==4.3.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsons==1.5.1
jsonschema==3.2.0
lazy-object-proxy==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
lxml==4.6.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mccabe==0.6.1
moment==0.12.1
numpy==1.19.3
opencensus-context==0.1.2
opencensus-ext-django==0.7.5
opencensus==0.7.13
packaging==21.0; python_version >= '3.6'
pandas==1.3.2
paramiko==2.7.2
pathspec==0.5.9
pdf417==0.8.1
pdfrw==0.4
pillow==8.3.1; python_version >= '3.6'
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycountry==20.7.3
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pylint==2.3.1
pynacl==1.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pypdf2==1.26.0
pypdf==1.13
pyrsistent==0.18.0; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.4
pytz==2021.1
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
regex==2021.8.28
reportlab==3.6.1
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.7.2; python_version >= '3.6'
semantic-version==2.8.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.14.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
svglib==1.1.0
termcolor==1.1.0
texttable==1.6.4
times==0.7
tinycss2==1.1.0; python_version >= '3.6'
typed-ast==1.4.3; implementation_name == 'cpython'
typish==1.9.3
tzlocal==3.0; python_version >= '3.6'
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.15.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
xlrd==2.0.1
zstandard==0.15.2; python_version >= '3.5'
